Add optional confidence parameter for cache
If confidence level of cache dictates that the client should
double check the cached response, an actual request to the
server will be performed even if a cached value exists.

If the cached value does not match the server response a call
to AbstractCache.feedback will be made and it is up to the cache
implementation on how to deal with that.