Refactor connection request/response flow
Refactor request/response flow to use Connection.create_request
which returns a future that can be resolved by the client without
holding on to the connection.

The actual response parsing is done in an independent task unless
a pubsub worker needs to explicitly fetch notifications that are not
based on an explicit request.