Remove re-creation of unused pool in cluster connection pool
The __default_node_queue was being recreated on every access
to the connection pool resulting in unncessary throwaway queues
being created. This also had the side effect of calls to release
resulting in needing a running event loop which should never be
necessary as a connection being released back to the pool must
have an associated pool.