Complete implementation of cluster transaction
Allow issuing watch & multi with a cluster pipeline as long
as the watched keys & the keys touched by the multi commands
are routed to the same node (They don't have to be on the same slot).

This change is a breaking change since the cluster  client's
transaction previously immediately moved the pipeline to multi mode
which would result in the callable not having to explicitly call multi
but also meant that the callable would not be able to fetch any data.