Change all redis command methods to return an awaitable
Instead of implementing each redis command as a coroutine
they are now regular methods that return an Awaitable.

This is (mostly) backward compatible with an advantage of allowing construction
of the command in a non async method and then being awaited in an async context.

Since the methods are no longer coroutines though, they can not be
passed to `asyncio.create_task` directly, and instead either need to be
wrapped in a coroutine, or wrapped into a task with `asyncio.ensure_future`