Refactor Pipeline to allow non async command stacking
Refactored pipeline to be in sync with the regular clients
and return awaitable CommandRequest objects from regular functions
instead of exposing coroutines only.

This allows pipelines to be constructed without using await and
as a side effect allows pipeline classes to be strict subclasses
of coredis.client.Client.

This additionally removes the complexity / overhead of maintaining
a type stub for the pipeline classes and also removes the dependency
on wrapt.