Ensure pubsub consumer exits when explicitly closed
Due to reconnection handling when a connection is lost
the pubsub consumer task would not break when cancelled
due to the underlying connection trapping the CancelledError
to be able to signal that there was a ConnectionError.

Instead of using a while True loop, stop processing when
initialized is set to False.